livebot/requirements.txt
MetaTrader5
pandas
numpy
ta
requests
vaderSentiment
textblob
python-dotenv
